src/worker/worker_video/requirements.txt
pika==1.2.0
moviepy==1.0.3
numpy==1.26.4
pillow==9.5.0
psycopg2-binary==2.9.3
sqlalchemy==1.4.31
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.1